% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobuffer-package.R
\docType{package}
\name{geobuffer-package}
\alias{geobuffer}
\alias{geobuffer-package}
\title{Geodesic buffer around points (long, lat) using metric radius}
\description{
Package that allows the possibility of creating geodesic buffers
when the radius is given in metric units. A geodesic buffer is not affected
by the distortions introduced by projected coordinate systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/valentinitnelav/geobuffer}
  \item Report bugs at \url{https://github.com/valentinitnelav/geobuffer/issues}
}

}
\author{
Valentin Stefan
}
\keyword{internal}
